refactor(data_migration): split CSV import/export into csv_transfer.rs
lib.rs was 5079 lines. Move the CSV-specific slice (export_to_csv,
import_goals_from_csv, their formula-injection helpers, and
CsvGoalRow) into a dedicated csv_transfer module -- a cohesive, mostly
self-contained unit with a clear boundary from the rest of the crate
(JSON/binary/encrypted-payload import-export, checksum/versioning,
migration tracking).

Behavior-preserving: every moved item is textually unchanged, just
relocated, and re-exported from the crate root
(`pub use csv_transfer::{export_to_csv, import_goals_from_csv}`) so
existing call sites and `use super::*;` test imports keep working
without modification.
